LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/GUM-invalid-facing-mode.https-expected.txt

Description

This test checks that trying to set an empty facingMode value in getUserMedia results in an OverconstrainedError.


FAIL Tests that setting an invalid facingMode constraint in getUserMedia fails assert_equals: expected "OverconstrainedError" but got "Error"
